uploading work-in-progress code for the Repository instantiation
In looking at ways to mock out the interaction with the Go-GitHub
client, I've found it to be onerous to just instantiate a Repository,
particularly because the fields are mostly pointers.

Anyway, this is the creation of a unit test, which sets up httptest in a way
to inject JSON at the right time and place to get the Repository instance,
and then inspect basic properties of the Repository.